third_party/blink/web_tests/dom/domparsing/serialize-cdata.html

<title>Test for "http://bugs.webkit.org/show_bug.cgi?id=11772":
XMLSerializer.serializeToString incorrect value for CDATA nodes.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
var d = window.document.implementation.createDocument("", "", null);
var n = d.createCDATASection("-- <cdata> --");
var s = (new XMLSerializer()).serializeToString(n);
assert_equals(s, "<![CDATA[-- <cdata> --]]>");
done();
</script>
